cleaning: move Statement code into new package
