+++ src/main/java/itemlease/Lease.java
@@ -20,4 +20,24 @@ public class Lease {
     public LeaseItem getGame() {
         return _leaseItem;
     }
+
+    public double getAmount() {
+        double amount = 0;
+        switch (this.getGame().getPriceCode()) {
+            case LeaseItem.REGULAR:
+                amount += 2;
+                if (this.getDaysLeased() > 2)
+                    amount += (this.getDaysLeased() - 2) * 1.5;
+                break;
+            case LeaseItem.NEWLY_RELEASED:
+                amount += this.getDaysLeased() * 3;
+                break;
+            case LeaseItem.CHILDREN:
+                amount += 1.5;
+                if (this.getDaysLeased() > 3)
+                    amount += (this.getDaysLeased() - 3) * 1.5;
+                break;
+        }
+        return amount;
+    }
 }
