+++ src/main/java/itemlease/statement/StatementFormatter.java
@@ -1,4 +1,6 @@
-package itemlease;
+package itemlease.statement;
+
+import itemlease.Customer;
 
 public interface StatementFormatter {
     String format(Customer customer);
