+++ src/main/java/itemlease/statement/StatementModel.java
@@ -1,4 +1,7 @@
-package itemlease;
+package itemlease.statement;
+
+import itemlease.Lease;
+import itemlease.LeaseItem;
 
 public class StatementModel {
     private double totalAmount;
@@ -25,28 +28,8 @@ public class StatementModel {
         this.nbLoyaltyPoints = nbLoyaltyPoints;
     }
 
-    static double leaseAmount(Lease lease) {
-        double amount = 0;
-        switch (lease.getGame().getPriceCode()) {
-            case LeaseItem.REGULAR:
-                amount += 2;
-                if (lease.getDaysLeased() > 2)
-                    amount += (lease.getDaysLeased() - 2) * 1.5;
-                break;
-            case LeaseItem.NEWLY_RELEASED:
-                amount += lease.getDaysLeased() * 3;
-                break;
-            case LeaseItem.CHILDREN:
-                amount += 1.5;
-                if (lease.getDaysLeased() > 3)
-                    amount += (lease.getDaysLeased() - 3) * 1.5;
-                break;
-        }
-        return amount;
-    }
-
     public double addLeaseAndReturnAmountAdded(Lease each) {
-        double amount = leaseAmount(each);
+        double amount = each.getAmount();
         this.setNbLoyaltyPoints(this.getNbLoyaltyPoints() + 1);
         if ((each.getGame().getPriceCode() == LeaseItem.NEWLY_RELEASED) && each.getDaysLeased() > 1)
             this.setNbLoyaltyPoints(this.getNbLoyaltyPoints() + 1);
